Features/crud detalle venta

+++ src/main/java/com/catalogo/dragon_ball/dto/DetalleVentaRequestDTO.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DetalleVentaRequestDTO {
-    private Integer idProducto;
-    private Integer cantidad;
+    private Long idProducto;
+    private Long cantidad;
+    
 }
